requirements.txt
absl-py==0.2.2
astor==0.7.1
backcall==0.1.0
certifi==2018.4.16
chardet==3.0.4
colorama==0.3.9
dask==0.18.1
decorator==4.3.0
future==0.16.0
gast==0.2.0
grpcio==1.13.0
gym==0.10.5
gym-retro==0.5.6
idna==2.7
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.1
jupyter-client==5.2.3
jupyter-core==4.4.0
Markdown==2.6.11
numpy==1.14.5
opencv-python==3.4.1.15
pandas==0.23.3
parso==0.3.1
pickleshare==0.7.4
Pillow==5.2.0
prompt-toolkit==1.0.15
protobuf==3.6.0
pyglet==1.3.2
Pygments==2.2.0
python-dateutil==2.7.3
pytz==2018.5
pyzmq==17.1.0
requests==2.19.1
simplegeneric==0.8.1
six==1.11.0
tensorboard==1.9.0
tensorflow-gpu==1.9.0
termcolor==1.1.0
tornado==5.1
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
Werkzeug==0.14.1
wincertstore==0.2
